Added -show-gui option
When using this tool its handy to be able to tell what files the program is stuck on. By using -show-gui you will get a very simple printout of all files the script is working on, and how long each file has been processed for.

There are also some changes related to argument parsing that were required for our setup.